+++ next/package.json
@@ -23,8 +23,8 @@
     "formidable": "^3.5.2",
     "next": "15.2.1",
     "node-fetch": "^2.7.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-quill": "^2.0.0"
   },
   "devDependencies": {
