+++ next/src/app/ClubRegister/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-// 引入必要的套件和服務
 import {
   Alert,
   Box,
@@ -11,59 +10,58 @@ import {
   TextField,
   Typography,
 } from "@mui/material";
-import {
-  createUserWithEmailAndPassword,
-  fetchSignInMethodsForEmail,
-} from "firebase/auth";
+import { createUserWithEmailAndPassword } from "firebase/auth";
 import Link from "next/link";
+import { useRouter } from 'next/navigation'; // ✅ 正確！
 import { ChangeEvent, FormEvent, useState } from "react";
-import styles from "../assets/globals.module.css";
-import { auth } from "../firebase/config";
-import { companyServices } from "../firebase/services";
+import styles from "../../assets/globals.module.css";
+import { auth } from "../../firebase/config";
+import { clubServices } from "../../firebase/services";
 
-// 簡化的表單資料類型，新增密碼欄位
+// Define types for formData and errors
 interface FormData {
-  companyName: string;
-  businessId: string;
+  clubName: string;
+  schoolName: string;
   email: string;
   password: string;
   confirmPassword: string;
 }
 
-// 簡化的表單錯誤類型，新增密碼錯誤
 interface FormErrors {
-  companyName: string;
-  businessId: string;
+  clubName: string;
+  schoolName: string;
   email: string;
   password: string;
   confirmPassword: string;
 }
 
-export default function CompanyRegister() {
-  // 表單狀態
+export default function ClubRegister() {
+  const router = useRouter();
+
+  // Form state
   const [formData, setFormData] = useState<FormData>({
-    companyName: "",
-    businessId: "",
+    clubName: "",
+    schoolName: "",
     email: "",
     password: "",
     confirmPassword: "",
   });
 
-  // 驗證錯誤
+  // Validation errors
   const [errors, setErrors] = useState<FormErrors>({
-    companyName: "",
-    businessId: "",
+    clubName: "",
+    schoolName: "",
     email: "",
     password: "",
     confirmPassword: "",
   });
 
-  // 提交狀態
+  // Form submission state
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [submitSuccess, setSubmitSuccess] = useState(false);
   const [submitError, setSubmitError] = useState<string | null>(null);
 
-  // 處理輸入變化
+  // Handle text input changes
   const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
     setFormData((prev) => ({
@@ -71,7 +69,7 @@ export default function CompanyRegister() {
       [name]: value,
     }));
 
-    // 清除錯誤
+    // Clear error when user starts typing
     if (errors[name as keyof FormErrors]) {
       setErrors((prev) => ({
         ...prev,
@@ -80,13 +78,7 @@ export default function CompanyRegister() {
     }
   };
 
-  // 驗證統一編號格式 - 8位數字
-  const validateBusinessId = (id: string) => {
-    const regex = /^\d{8}$/;
-    return regex.test(id);
-  };
-
-  // 驗證電子郵件格式
+  // Validate email format
   const validateEmail = (email: string) => {
     const regex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     return regex.test(email);
@@ -98,22 +90,19 @@ export default function CompanyRegister() {
     return regex.test(password);
   };
 
-  // 表單驗證
+  // Form validation
   const validateForm = () => {
     let valid = true;
     const newErrors = { ...errors };
 
-    // 必填欄位驗證
-    if (!formData.companyName.trim()) {
-      newErrors.companyName = "此欄位為必填";
+    // Required fields validation
+    if (!formData.clubName.trim()) {
+      newErrors.clubName = "此欄位為必填";
       valid = false;
     }
 
-    if (!formData.businessId.trim()) {
-      newErrors.businessId = "此欄位為必填";
-      valid = false;
-    } else if (!validateBusinessId(formData.businessId)) {
-      newErrors.businessId = "請輸入有效的統一編號";
+    if (!formData.schoolName.trim()) {
+      newErrors.schoolName = "此欄位為必填";
       valid = false;
     }
 
@@ -147,11 +136,11 @@ export default function CompanyRegister() {
     return valid;
   };
 
-  // 處理表單提交
+  // Handle form submission
   const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
 
-    // 驗證表單
+    // Validate form
     if (!validateForm()) {
       return;
     }
@@ -160,64 +149,42 @@ export default function CompanyRegister() {
     setSubmitError(null);
 
     try {
-      // 先檢查電子郵件是否已註冊
-      const signInMethods = await fetchSignInMethodsForEmail(
-        auth,
-        formData.email
-      );
-      if (signInMethods.length > 0) {
-        throw new Error("auth/email-already-in-use");
-      }
-
-      // 使用電子郵件和用戶設定的密碼創建 Firebase Auth 用戶
+      // 1. 建立Firebase用戶
       const userCredential = await createUserWithEmailAndPassword(
         auth,
         formData.email,
         formData.password
       );
 
-      const userId = userCredential.user.uid;
-      console.log("用戶創建成功:", userId);
-
-      // 將簡化的公司資料存入 Firestore
-      const companyData = {
-        companyName: formData.companyName,
-        businessId: formData.businessId,
-        industryType: "其他", // 預設值
-        contactName: "", // 空值
-        contactPhone: "", // 空值
+      // 2. 將社團資料保存到Firestore
+      const clubData = {
+        clubName: formData.clubName,
+        schoolName: formData.schoolName,
         email: formData.email,
-        companyDescription: "", // 空值
-        cooperationFields: [], // 空值
-        logoURL: "", // 改為空字串而非 null
-        businessCertificateURL: "", // 改為空字串而非 null
-        status: "pending", // 預設狀態為待審核
+        status: "pending", // 初始狀態為待審核
         registrationDate: new Date().toISOString(),
-        userId: userId,
+        userId: userCredential.user.uid, // 將Firebase用戶ID關聯到社團資料
       };
 
-      // 存儲公司資料
-      const companyId = await companyServices.addCompany(companyData);
-      console.log("公司資料存儲成功:", companyId);
+      // 3. 添加到Firestore
+      await clubServices.addClub(clubData);
 
+      // 4. 註冊成功
       setSubmitSuccess(true);
-    } catch (error) {
-      console.error("註冊公司過程中發生錯誤:", error);
+    } catch (error: any) {
+      console.error("註冊錯誤:", error);
 
-      // 處理 Firebase 錯誤
-      let errorMessage = "註冊過程中發生錯誤，請稍後再試";
+      // 處理常見Firebase錯誤
+      let errorMessage = "註冊失敗，請稍後再試";
 
-      if (error instanceof Error) {
-        // 處理常見的 Firebase 註冊錯誤
-        if (error.message.includes("email-already-in-use")) {
-          errorMessage = "此電子郵件已被註冊，請使用其他電子郵件或嘗試登入";
-        } else if (error.message.includes("invalid-email")) {
-          errorMessage = "請提供有效的電子郵件地址";
-        } else if (error.message.includes("weak-password")) {
-          errorMessage = "密碼強度不足，請使用更複雜的密碼";
-        } else {
-          errorMessage = `註冊失敗: ${error.message}`;
-        }
+      if (error.code === "auth/email-already-in-use") {
+        errorMessage = "此電子郵件已被註冊，請使用其他電子郵件";
+      } else if (error.code === "auth/invalid-email") {
+        errorMessage = "請輸入有效的電子郵件地址";
+      } else if (error.code === "auth/weak-password") {
+        errorMessage = "密碼強度不足，請設置更強的密碼";
+      } else if (error.code === "auth/network-request-failed") {
+        errorMessage = "網絡連接失敗，請檢查您的網絡連接";
       }
 
       setSubmitError(errorMessage);
@@ -235,29 +202,29 @@ export default function CompanyRegister() {
               註冊成功！
             </Typography>
             <Typography align="center" sx={{ mb: 2 }}>
-              感謝您完成企業註冊，您的帳號已成功建立，我們將審核您的資料
+              感謝您完成社團註冊，您的帳號已成功建立
             </Typography>
             <Box sx={{ display: "flex", justifyContent: "center", mt: 3 }}>
               <Button
                 variant="contained"
                 component={Link}
-                href="/CompanyList"
+                href="/LoginPage"
                 sx={{ mr: 2 }}
               >
-                查看公司列表
-              </Button>
-              <Button variant="outlined" component={Link} href="/LoginPage">
                 前往登入
+              </Button>
+              <Button variant="outlined" component={Link} href="/">
+                返回首頁
               </Button>
             </Box>
           </Paper>
         ) : (
           <Paper elevation={3} sx={{ p: 4, mt: 4, mb: 4 }}>
             <Typography variant="h4" align="center" gutterBottom>
-              企業帳號註冊
+              社團帳號註冊
             </Typography>
             <Typography color="textSecondary" align="center" sx={{ mb: 2 }}>
-              請填寫基本資料以註冊企業帳號
+              請填寫基本資料以註冊社團帳號
             </Typography>
 
             {submitError && (
@@ -270,33 +237,28 @@ export default function CompanyRegister() {
               <Box
                 sx={{ display: "flex", flexDirection: "column", gap: 3, mt: 3 }}
               >
-                {/* 企業名稱 */}
+                {/* 社團名稱 */}
                 <TextField
-                  name="companyName"
-                  label="企業名稱"
-                  value={formData.companyName}
+                  name="clubName"
+                  label="社團名稱"
+                  value={formData.clubName}
                   onChange={handleInputChange}
-                  error={!!errors.companyName}
-                  helperText={errors.companyName}
+                  error={!!errors.clubName}
+                  helperText={errors.clubName}
                   fullWidth
                   required
                 />
 
-                {/* 統一編號 */}
+                {/* 學校名稱 */}
                 <TextField
-                  name="businessId"
-                  label="統一編號"
-                  value={formData.businessId}
+                  name="schoolName"
+                  label="學校名稱"
+                  value={formData.schoolName}
                   onChange={handleInputChange}
-                  error={!!errors.businessId}
-                  helperText={errors.businessId || "請輸入8位數字的統一編號"}
+                  error={!!errors.schoolName}
+                  helperText={errors.schoolName}
                   fullWidth
                   required
-                  sx={{
-                    "& .MuiInputBase-input": {
-                      maxLength: 8,
-                    },
-                  }}
                 />
 
                 {/* 電子郵件 */}
